loaddeck 1.8

+++ card.h
@@ -5,7 +5,7 @@
 class Card {
 protected:
     std::string name;
-    int cost;
+    size_t cost;
     virtual void readInfo(std::string name) = 0;
 
 public:
@@ -15,7 +15,7 @@ public:
     // virtual void play();
 
     std::string getName();
-    int getCost();
+    size_t getCost();
 };
 
 #endif
